+++ crates/tad-gui/src/helpers.rs
@@ -321,6 +321,39 @@ impl InputForm {
         (w1, w2)
     }
 
+    pub fn add_three_inputs<T, U, V>(
+        &mut self,
+        text: &str,
+        ch_units1: i32,
+        ch_units2: i32,
+    ) -> (T, U, V)
+    where
+        T: WidgetExt + Default,
+        U: WidgetExt + Default,
+        V: WidgetExt + Default,
+    {
+        let w1_width = ch_units_to_width(&self.group, ch_units1);
+        let w2_width = ch_units_to_width(&self.group, ch_units2);
+
+        let mut r = Flex::default().row();
+        self.group.fixed(&r, self.row_height);
+
+        let l = label(text);
+        r.fixed(&l, self.left_column_width);
+
+        let w1 = T::default();
+        r.fixed(&w1, w1_width);
+
+        let w2 = U::default();
+        r.fixed(&w2, w2_width);
+
+        let w3 = V::default();
+
+        r.end();
+
+        (w1, w2, w3)
+    }
+
     pub fn add_two_inputs_right<T, U>(&mut self, text: &str, ch_units: i32) -> (T, U)
     where
         T: WidgetExt + Default,