+++ docs/changelog.md
@@ -10,6 +10,7 @@ Version 0.0.4
  * Added `s` play sample MML command
  * Fixed `n` play MIDI note MML command erroneously parsing note length
  * Added `$` hexadecimal number parsing
+ * Added loop point filter overrides to the BRR encoder
 
 
 Version 0.0.3